+++ tools/sertxt_test/main.cpp
@@ -94,12 +94,12 @@ static void TestSettingsDeserialize()
         printf("failed to load '%s'", path);
         return;
     }
-    Settings *settings = DeserializeSettings(s, (int)fileSize);
+    Settings *settings = DeserializeSettings(s, fileSize);
     if (!settings) {
         printf("failed to deserialize\n'%s'\n", s);
         return;
     }
-    int serializedLen;
+    size_t serializedLen;
     char *s2 = (char*)SerializeSettings(settings, &serializedLen);
     s2 += 3; // skip utf8 bom
     char *toFree = s;
